+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 10.0
 - Remove semicolon before "With" & "On" values in PrintRequests.php
 - HTML Link is selected: bold in ScheduleReport.php
 - Display Period title if no short name set in IncompleteSchedules.php
+- Fix Widget search & add Search Terms header in IncompleteSchedules.php
 
 Changes in 9.3
 --------------